+++ src/main/java/com/example/kuhidbs/dto/company/재무/CFncDTO.java
@@ -2,6 +2,8 @@ package com.example.kuhidbs.dto.company.재무;
 
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @Data
@@ -13,12 +15,12 @@ public class CFncDTO {
     private String companyId; // Company 객체 대신 ID만 포함
     private Integer financialYear;
     private String financialHalf;
-    private Integer revenue;
-    private Integer operatingProfit;
-    private Integer netIncome;
-    private Integer totalAssets;
-    private Integer totalCapital;
-    private Integer capital;
+    private BigDecimal revenue;
+    private BigDecimal operatingProfit;
+    private BigDecimal netIncome;
+    private BigDecimal totalAssets;
+    private BigDecimal totalCapital;
+    private BigDecimal capital;
     private Integer employeeCount;
-    private Integer totalDebt;
+    private BigDecimal totalDebt;
 }
